requirements.txt
numpy==1.26.4

torch==2.9.1
transformers==4.57.1

PyPDF2==3.0.1

faiss-cpu==1.12.0

requests==2.32.5
openai==2.8.0

tqdm==4.67.1
